email-worker/requirements.txt
pymongo
sendgrid
python-dotenv
pytest
pytest-cov
